stdout formatting
