+++ webapp/common/logging/stdout_handler.py
@@ -6,9 +6,11 @@ Use of this source code is governed by an MIT-style license that can be found in
 import logging
 from logging import StreamHandler
 
+from .otel_formatter import OtelFormatter
+
 
 class StdoutHandler(StreamHandler):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.setLevel(logging.DEBUG)
-        self.setFormatter(logging.Formatter('%(asctime)s %(levelname)s %(tags)s: %(message)s '))
+        self.setFormatter(OtelFormatter())
